% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobstatus_classes.R
\docType{data}
\name{jobstatus}
\alias{jobstatus}
\title{jobstatus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
jobstatus
}
\description{
create a job status node to record progress on a task
}
\keyword{datasets}
