% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-future.R
\name{subjob_future}
\alias{subjob_future}
\alias{resolved.Subjob.Future}
\alias{result.Subjob.Future}
\title{subjob_future}
\usage{
subjob_future(expr, envir = parent.frame(), substitute = TRUE,
  globals = TRUE, packages = NULL, lazy = FALSE, seed = NULL,
  evaluator = plan("next"), ...)

\method{resolved}{Subjob.Future}(x, timeout = 0.2, ...)

\method{result}{Subjob.Future}(future, ...)
}
\description{
a drop-in replacement for \code{\link[future:future]{future::future}} that passes on job status
information
}
