% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_jobstatus.R
\name{with_jobstatus}
\alias{with_jobstatus}
\title{with_jobstatus}
\usage{
with_jobstatus(expr, display = jobstatus_bar())
}
\arguments{
\item{expr}{an expression to execute}

\item{display}{a function to read status information and display the progress
information}
}
\description{
run an expression with some jobs, transmitting and possibly displaying status
information from jobstatus objects in the jobs
}
